add a separate file for graphql types
